requirements.txt
streamlit>=1.28.0
groq>=0.4.0
pillow>=10.0.0
fpdf>=1.7.2
